Abhay/clip_and_whisper_tests/requirements.txt
--index-url https://download.pytorch.org/whl/cu128
torch==2.8.0+cu128
torchvision
torchaudio
openai-whisper
open_clip_torch
ffmpeg-python
tqdm
numpy